MAINT: refactor tests
Move tests of projections to a new file. This is motivated because
it is more useful to have tests of the same category of behavior
in the same file, and the original file polytope_test.py is now so
large that it is difficult to navigate.
